+++ NEWS.md
@@ -8,6 +8,7 @@
 * Better handling of larger datasets when running `tabnet_explain`.
 * Batch predictions to avoid OOM error.
 * Add tabnet_pretrain() for unsupervised pretraining (@cregouby, #29 )
+* Add autoplot() of model loss among epochs (@cregouby, #34)
 
 # tabnet 0.1.0